Fixes for Xcode

+++ source/types.inl
@@ -17,20 +17,20 @@ enum EFileType {
 struct cc_file_t_ {
   enum EFileType file_type;
   const char* path;
-  unsigned parent_group_idx;
+  size_t parent_group_idx;
 };
 
 struct cc_file_custom_command_t_ {
   enum EFileType file_type;
   const char* path;
-  unsigned parent_group_idx;
+  size_t parent_group_idx;
   const char* output_file;
   const char* command;
 };
 
 typedef struct cc_group_impl_t {
   const char* name;
-  unsigned parent_group_idx;
+  size_t parent_group_idx;
 } cc_group_impl_t;
 
 typedef struct cc_state_impl_t {
@@ -55,7 +55,7 @@ typedef struct cc_project_impl_t {
   cc_platform_impl_t** platforms;    /* stretch array */
   const char* postBuildAction;
 
-  unsigned parent_group_idx;
+  size_t parent_group_idx;
 } cc_project_impl_t;
 
 struct {
@@ -72,9 +72,9 @@ struct {
 cc_group_t cc_group_create(const char* in_group_name, const cc_group_t in_parent_group) {
   cc_group_impl_t group;
   group.name             = cc_printf("%s", in_group_name);
-  group.parent_group_idx = (unsigned)in_parent_group;
+  group.parent_group_idx = (size_t)in_parent_group;
   array_push(cc_data_.groups, group);
-  return (cc_group_t)(array_count(cc_data_.groups) - 1);
+  return (cc_group_t)(size_t)(array_count(cc_data_.groups) - 1);
 }
 
 cc_state_t cc_state_create() {
@@ -107,7 +107,7 @@ cc_project_t cc_project_create_(const char* in_project_name, EProjectType in_pro
   p->name                = name_copy;
   array_push(cc_data_.projects, p);
 
-  p->parent_group_idx = (unsigned)in_parent_group;
+  p->parent_group_idx = (size_t)in_parent_group;
   return p;
 }
 
@@ -123,7 +123,7 @@ void addFilesToProject(cc_project_t in_out_project, unsigned num_files,
     struct cc_file_t_* file_data = (struct cc_file_t_*)cc_alloc_(sizeof(struct cc_file_t_));
     file_data->file_type         = FileTypeCustomCommand;
     file_data->path              = cc_printf("%s", *in_file_names);
-    file_data->parent_group_idx  = (unsigned)in_parent_group;
+    file_data->parent_group_idx  = (size_t)in_parent_group;
     array_push(project->file_data, file_data);
   }
 }
@@ -139,7 +139,7 @@ void addFilesFromFolderToProject(cc_project_t in_out_project, const char* folder
     struct cc_file_t_* file_data = (struct cc_file_t_*)cc_alloc_(sizeof(struct cc_file_t_));
     file_data->file_type         = FileTypeCustomCommand;
     file_data->path              = file_path;
-    file_data->parent_group_idx  = (unsigned)in_parent_group;
+    file_data->parent_group_idx  = (size_t)in_parent_group;
     array_push(project->file_data, file_data);
   }
 }
@@ -154,7 +154,7 @@ void cc_project_addFileWithCustomCommand(cc_project_t in_out_project, const char
       (struct cc_file_custom_command_t_*)cc_alloc_(sizeof(struct cc_file_custom_command_t_));
   file_data->file_type        = FileTypeCustomCommand;
   file_data->path             = cc_printf("%s", in_file_name);
-  file_data->parent_group_idx = (unsigned)in_parent_group;
+  file_data->parent_group_idx = (size_t)in_parent_group;
   file_data->command          = cc_printf("%s", in_custom_command);
   file_data->output_file      = cc_printf("%s", in_output_file_name);
   array_push(project->file_data_custom_command, file_data);
